Remove a confusing line from the README
It seems that distinguishing between individual test FILE and individual
TEST is not useful. Running an individual test FILE is documented above
this, and running an individual TEST *is* supported with minitest's -n
flag, which is kind of annoying but people who want to do that can
search it out on their own.