Update to use v2.1, add shim to handle polymorphism on event types
This change updates the library to be able to access data on the v2.1
API, including extension fields on certain types of records. The code
was regenerated from the Swagger definitions on Clever's repo, but in
testing we found what looked like custom changes (or at least
off-instruction changes to the API client) to support rich data for
Clever Events).

We generated the code using Swagger codegen:

```
swagger-codegen generate -i swagger-api/v2.1-client.yml \
  -l ruby -o clever-ruby --additional-properties \
  gemName=clever-ruby,moduleName=Clever,gemVersion=2.1.0
```

But after that, we found that access to the object payload for Clever
Events had disappeared due to the client not converting Clever events
into their subtype based on the `discriminator` field. As a result,
events had high-level data, but none of the deep payload contents needed
to avoid data fetches.

Swagger allows for type inheritance, but there are some known gaps in
the implementation of clients to handle this in some languages. This
isn't part of the documented steps to generate new API documentation,
but without it, it's not possible to consume events from the Event API
client with their payloads (they all come back as generic events with
ID, record type, and created at date with the rest of the API data lost
in conversion to `Clever::Event`). This change isn't documented in the
README steps to generate a client, but adding it back in from the code
lost after regenerating definitions from Swagger allows us to operate
successfully on Events data again.